+++ config/console/bootstrappers.php
@@ -1,7 +1,7 @@
 <?php
-use Opulence\Framework\Bootstrappers\Console\Commands\CommandsBootstrapper as OpulenceCommandsBootstrapper;
-use Opulence\Framework\Bootstrappers\Console\Requests\RequestsBootstrapper;
-use Opulence\Framework\Bootstrappers\Console\Composer\ComposerBootstrapper;
+use Opulence\Framework\Composer\Bootstrappers\ComposerBootstrapper;
+use Opulence\Framework\Console\Bootstrappers\CommandsBootstrapper as OpulenceCommandsBootstrapper;
+use Opulence\Framework\Console\Bootstrappers\RequestBootstrapper;
 use Project\Application\Bootstrappers\Console\Commands\CommandsBootstrapper as ProjectCommandsBootstrapper;
 use Project\Application\Bootstrappers\Http\Routing\RouterBootstrapper;
 use Project\Application\Bootstrappers\Http\Views\ViewBootstrapper;
@@ -14,7 +14,7 @@ use Project\Application\Bootstrappers\Http\Views\ViewBootstrapper;
 return [
     RouterBootstrapper::class,
     OpulenceCommandsBootstrapper::class,
-    RequestsBootstrapper::class,
+    RequestBootstrapper::class,
     ComposerBootstrapper::class,
     ViewBootstrapper::class,
     ProjectCommandsBootstrapper::class