+++ tests/src/Project/Application/Console/IntegrationTestCase.php
@@ -2,7 +2,7 @@
 namespace Project\Application\Console;
 
 use Opulence\Bootstrappers\ApplicationBinder;
-use Opulence\Framework\Testing\PhpUnit\Console\IntegrationTestCase as BaseIntegrationTestCase;
+use Opulence\Framework\Console\Testing\PhpUnit\IntegrationTestCase as BaseIntegrationTestCase;
 use Opulence\Ioc\IContainer;
 
 /**